split date filter into style wrapper and logic component
